% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temperature_loss.R
\name{air2water}
\alias{air2water}
\title{Estimate air temperature from air temperature}
\usage{
air2water(Tair)
}
\arguments{
\item{Tair}{air temperature [degrees C]}
}
\value{
estimated ater temperature [degrees C]
}
\description{
Estimate air temperature from air temperature
}
