% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_integration_period.R
\name{basin_integration_period}
\alias{basin_integration_period}
\title{Select an integration period for a basin}
\usage{
basin_integration_period(
  upstream_capacity,
  expected_monthly_flow,
  available_periods
)
}
\arguments{
\item{upstream_capacity}{upstream storage capacity available for electric
power generation}

\item{expected_monthly_flow}{typical monthly total blue water, e.g., the location
parameter of a fitted distribution}

\item{available_periods}{vector of available integration periods in months}
}
\description{
Select an integration period for a basin
}
