% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_upstream_capacity.R
\name{basin_upstream_capacity}
\alias{basin_upstream_capacity}
\title{Assign an upstream capacity to each basin}
\usage{
basin_upstream_capacity(basins, dams)
}
\arguments{
\item{basins}{a \code{sf} with a field for id 
(\code{basin_id}), downstream
basin id (\code{downstream_id})}

\item{dams}{a \code{sf} with a field for
capacity (\code{capacity})}
}
\value{
a data frame with fields \code{hybas_id}
              and \code{months_storage}
}
\description{
Assign an upstream capacity to each basin
}
