% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydropower_loss.R
\name{hydropower_loss}
\alias{hydropower_loss}
\title{Estimate hydropowerloss based on blue water}
\usage{
hydropower_loss(blue_water, blue_water_median)
}
\arguments{
\item{blue_water}{observed blue water}

\item{blue_water_median}{median blue water}
}
\value{
lost fraction of expected hydropower generation
}
\description{
Estimate hydropowerloss based on blue water
}
