% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_losses.R
\name{summarize_losses}
\alias{summarize_losses}
\title{Summarize aggregated loss risks for multiple plants in a given boundary}
\usage{
summarize_losses(plants, loss, aggfield, hours_in_period)
}
\arguments{
\item{plants}{data frame of plants}

\item{loss}{data frame of losses}

\item{aggfield}{field over which to aggregate}

\item{hours_in_period}{number of hours, to convert power (MW) to energy (MW-h)}
}
\value{
data frame with aggregated risks
}
\description{
Summarize aggregated loss risks for multiple plants in a given boundary
}
