% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temperature_loss.R
\name{temperature_loss}
\alias{temperature_loss}
\title{Temperature loss}
\usage{
temperature_loss(
  To = NA,
  To_rp = NA,
  Tbas = NA,
  Tc = NA,
  Tc_rp = NA,
  Teff = NA,
  eff = 0.005,
  Treg = NA,
  Tdiff = NA
)
}
\arguments{
\item{To}{air temperature at plant [degrees C]}

\item{To_rp}{air temperature at plant as a return period (yr). Negative values
indicate cold anomalies; positive values indicate warm
anomalies.}

\item{Tbas}{basin air temperature [degrees C]}

\item{Tc}{optional air temperature [degrees C] at which cold losses begin}

\item{Tc_rp}{return period at which cold losses begin}

\item{Teff}{optional air temperature [degrees C] at which efficiency loss begins}

\item{eff}{efficiency loss per degree C above Teff}

\item{Treg}{optional regulatory limit water temperature [degrees C]}

\item{Tdiff}{effluent - influent water temperature [degrees C]}
}
\description{
Temperature loss
}
