% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_temperature_loss.R
\name{water2air}
\alias{water2air}
\title{Estimate water temperature from air temperature}
\usage{
water2air(Twater)
}
\arguments{
\item{Twater}{water temperature [degrees C]}
}
\value{
estimated air temperature [degrees C]
}
\description{
Estimate water temperature from air temperature
}
