% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_cooled_loss.R
\name{water_cooled_loss}
\alias{water_cooled_loss}
\title{Estimate water-cooled loss risk}
\usage{
water_cooled_loss(x, xc, xmax)
}
\arguments{
\item{x}{total blue water return period}

\item{xc}{return period at which loss risk begins to occur}

\item{xmax}{return period associated with complete loss risk}
}
\description{
Estimate water-cooled loss risk
}
