% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_cooled_loss.R
\name{water_cooled_loss_onset}
\alias{water_cooled_loss_onset}
\title{Estimate onset of water-cooled loss}
\usage{
water_cooled_loss_onset(baseline_water_stress)
}
\arguments{
\item{baseline_water_stress}{baseline water stress}
}
\value{
return period at which losses begin
}
\description{
Estimate onset of water-cooled loss
}
